requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
tqdm==4.66.4
requests==2.32.4
pillow==10.4.0
numpy==1.26.4
torch==2.7.1+cu118
torchvision==0.22.1+cu118
chromadb==0.4.24
einops==0.8.0
python-doctr==1.0.0
PyYAML==6.0.1
clip-benchmark==1.6.2
datasets==4.0.0
open-clip-torch==3.0.0
timm==1.0.19
transformers==4.53.3
tensorflow==2.19.0
Flask==3.1.1
flask-cors==6.0.1
ttkbootstrap==1.14.1
darkdetect==0.8.0
opencv-python==4.11.0.86
mtcnn-opencv==1.0.2
deepface==0.0.93
flowlauncher==0.2.0
rank-bm25==0.2.2
ftfy==6.3.1
regex==2024.11.6
scikit-learn==1.7.1
pydantic-settings==2.7.1
python-dotenv==1.0.1
google-auth-oauthlib==1.2.2
google-auth==2.38.0
google-api-python-client==2.163.0
gunicorn==23.0.0
